Add cash/UPI separation, host gift acceptance, and WhatsApp single-tap UX
- Cash/UPI split: helpers and host see cash vs UPI collected separately;
  helpers cannot spend UPI money (only cash)
- Host collect: new HostCollectScreen + /api/app/host/collect endpoint
  so host can record in-person gifts directly
- WhatsApp menu redesign: multi-section list (QR_N/EDIT_N/HIST_N) lets
  user pick action + person in a single tap instead of two round-trips
- Single registration message: combines success text with menu in one list msg
- Decimal amounts: V16 migration + ddl-auto=none to fix DECIMAL/float mismatch
- Flyway: ignore-missing-migrations=true for V11-V15 applied on DB only

+++ chanlo/src/main/java/com/mysteriousmonkeys/chanlo/dashboard/GuestHistoryController.java
@@ -113,7 +113,7 @@ public class GuestHistoryController {
             "persons", persons,
             "totalAmount", payments.stream()
                 .filter(h -> h.getPaymentStatus() == com.mysteriousmonkeys.chanlo.money.PaymentStatus.SUCCESS)
-                .mapToLong(Hisab::getAmount)
+                .mapToDouble(Hisab::getAmount)
                 .sum()
         ));
     }
